requirements.txt
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
distlib==0.3.7
docutils==0.20.1
filelock==3.12.3
identify==2.5.27
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
nodeenv==1.8.0
packaging==23.1
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.0.1
pycparser==2.21
Pygments==2.16.1
pypandoc==1.11
pyproject-api==1.5.4
PyYAML==6.0.1
readme-renderer==41.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
SecretStorage==3.3.3
six==1.16.0
svgwrite==1.4.3
tox==4.10.0
Tree==0.2.4
twine==4.0.2
urllib3==2.0.4
virtualenv==20.24.3
webencodings==0.5.1
zipp==3.16.2
